Start of a no vector implementation of standard.

+++ standard/interfaces/shape.hpp
@@ -4,7 +4,7 @@ bool Shape<T>::isInShape(const Coordinate<T>& p) const
   const auto p_xy = static_cast<const Cartesian<T>&>(p);
   bool ret = false;
 
-  for(int i=0, j=static_cast<long>(numSides())-1; i<numSides(); j=i++)
+  for(size_t i=0, j=numSides()-1; i<numSides(); j=i++)
   {
     const auto& verti = (*this)[i];
     const auto& vertj = (*this)[j];
